# docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Araponga API",
    "version": "v1",
    "description": "Araponga é uma plataforma comunitária orientada a território. Esta API expõe recursos de território, feed comunitário, entidades do mapa e saúde do território. O objetivo do MVP é viabilizar cadastro, visualização e curadoria comunitária com segurança e governança mínima.",
    "contact": {
      "name": "Araponga (maintainers)"
    },
    "license": {
      "name": "MIT"
    }
  },
  "paths": {
...
    "/api/v1/stores": {
      "post": {
        "tags": [
          "Stores"
        ],
        "summary": "Cria ou atualiza a loja do usuário no território.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertStoreRequest"
              },
              "examples": {
                "createStore": {
                  "summary": "Criar loja",
                  "value": {
                    "territoryId": "22222222-2222-2222-2222-222222222222",
                    "displayName": "Loja do Vale",
                    "description": "Produtos artesanais da comunidade.",
                    "contactVisibility": "ON_INQUIRY_ONLY",
                    "contact": {
                      "phone": "(11) 90000-0000",
                      "whatsapp": "(11) 90000-0000",
                      "email": "loja@comunidade.com",
                      "instagram": "@lojavale",
                      "website": "https://lojavale.com",
                      "preferredContactMethod": "whatsapp"
                    }
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoreResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/v1/stores/me": {
      "get": {
        "tags": [
          "Stores"
        ],
        "summary": "Retorna a loja do usuário no território.",
        "parameters": [
          {
            "name": "territoryId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoreResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/v1/stores/{id}": {
      "patch": {
        "tags": [
          "Stores"
        ],
        "summary": "Atualiza detalhes da loja.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStoreRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoreResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/v1/stores/{id}/pause": {
      "post": {
        "tags": [
          "Stores"
        ],
        "summary": "Altera status da loja para PAUSED.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoreResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/v1/stores/{id}/activate": {
      "post": {
        "tags": [
          "Stores"
        ],
        "summary": "Altera status da loja para ACTIVE.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoreResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/v1/stores/{id}/archive": {
      "post": {
        "tags": [
          "Stores"
        ],
        "summary": "Altera status da loja para ARCHIVED.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoreResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/v1/stores/{id}/payments/enable": {
      "post": {
        "tags": [
          "Stores"
        ],
        "summary": "Define se a loja aceita pagamentos.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StorePaymentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoreResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "StoreContactPayload": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "nullable": true
          },
          "whatsapp": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "instagram": {
            "type": "string",
            "nullable": true
          },
          "website": {
            "type": "string",
            "nullable": true
          },
          "preferredContactMethod": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "UpsertStoreRequest": {
        "required": [
          "territoryId",
          "displayName",
          "contactVisibility"
        ],
        "type": "object",
        "properties": {
          "territoryId": {
            "type": "string",
            "format": "uuid"
          },
          "displayName": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "contactVisibility": {
            "type": "string"
          },
          "contact": {
            "$ref": "#/components/schemas/StoreContactPayload"
          }
        }
      },
      "UpdateStoreRequest": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "contactVisibility": {
            "type": "string",
            "nullable": true
          },
          "contact": {
            "$ref": "#/components/schemas/StoreContactPayload"
          }
        }
      },
      "StorePaymentRequest": {
        "required": [
          "enabled"
        ],
        "type": "object",
        "properties": {
          "enabled": {
            "type": "boolean"
          }
        }
      },
      "StoreResponse": {
        "required": [
          "id",
          "territoryId",
          "ownerUserId",
          "displayName",
          "status",
          "paymentsEnabled",
          "contactVisibility",
          "createdAtUtc",
          "updatedAtUtc"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "territoryId": {
            "type": "string",
            "format": "uuid"
          },
          "ownerUserId": {
            "type": "string",
            "format": "uuid"
          },
          "displayName": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string"
          },
          "paymentsEnabled": {
            "type": "boolean"
          },
          "contactVisibility": {
            "type": "string"
          },
          "contact": {
            "$ref": "#/components/schemas/StoreContactPayload"
          },
          "createdAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAtUtc": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  }
}